Update: - Refactor Controller layer. - Change Url into ImageData in naming convention

+++ Backend_Ecommerce/Ecommerce.Controller/src/Controller/CategoryController.cs
@@ -3,54 +3,81 @@ using Ecommerce.Service.src.ServiceAbstract;
 using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
 
-
 namespace Ecommerce.Controller.src.Controller
 {
     [ApiController]
     [Route("api/v1/categories")]
     public class CategoryController : ControllerBase
     {
+        #region Properties
         private readonly ICategoryService _categoryService;
+        #endregion
+
+        #region Constructor
         public CategoryController(ICategoryService categoryService)
         {
             _categoryService = categoryService;
         }
+        #endregion
 
+        #region GET http://localhost:5227/api/v1/categories
         [HttpGet]
-        public async Task<IEnumerable<CategoryReadDto>> GetAllCategoriesAsync() // endpoint: /categories
+        public async Task<ActionResult<IEnumerable<CategoryReadDto>>> GetAllCategoriesAsync()
         {
-
-            return await _categoryService.GetAllCategoriesAsync();
-
+            var result = await _categoryService.GetAllCategoriesAsync();
+            return Ok(result);
         }
+        #endregion
 
-        [HttpGet("{categoryId}")] // endpoint: /categories/:category_id
-        public async Task<CategoryReadDto> GetCategoryByIdAsync([FromRoute] Guid categoryId)
+        #region GET http://localhost:5227/api/v1/categories/:category_id
+        [HttpGet("{categoryId:guid}")]
+        public async Task<ActionResult<CategoryReadDto>> GetCategoryByIdAsync([FromRoute] Guid categoryId)
         {
-            return await _categoryService.GetCategoryByIdAsync(categoryId);
+            var result = await _categoryService.GetCategoryByIdAsync(categoryId);
+            return Ok(result);
         }
+        #endregion
 
+        #region POST http://localhost:5227/api/v1/categories
         [Authorize(Roles = "Admin")]
         [HttpPost] // endpoint: /categories
-        public async Task<CategoryReadDto> CreateCategoryAsync([FromBody] CategoryCreateDto categoryCreateDto)
+        public async Task<ActionResult<CategoryReadDto>> CreateCategoryAsync([FromBody] CategoryCreateDto categoryCreateDto)
         {
-            return await _categoryService.CreateCategoryAsync(categoryCreateDto);
-
+            var result = await _categoryService.CreateCategoryAsync(categoryCreateDto);
+            return Ok(result);
         }
+        #endregion
 
+        #region PUT http://localhost:5227/api/v1/categories/:category_id
         [Authorize(Roles = "Admin")]
-        [HttpPut("{categoryId}")] // endpoint: /categories/:category_id
-        public async Task<CategoryReadDto> UpdateCategoryByIdAsync([FromRoute] Guid categoryId, [FromBody] CategoryUpdateDto categoryUpdateDto)
+        [HttpPut("{categoryId}")]
+        public async Task<ActionResult<CategoryReadDto>> UpdateCategoryByIdAsync([FromRoute] Guid categoryId, [FromBody] CategoryUpdateDto categoryUpdateDto)
         {
-            return await _categoryService.UpdateCategoryByIdAsync(categoryId, categoryUpdateDto);
+            var foundCategory = await GetCategoryByIdAsync(categoryId);
+            if (foundCategory is null)
+            {
+                return NotFound("Category not found");
+            }
 
+            var result = await _categoryService.UpdateCategoryByIdAsync(categoryId, categoryUpdateDto);
+            return Ok(result);
         }
+        #endregion
 
+        #region DELETE http://localhost:5227/api/v1/categories/:category_id
         [Authorize(Roles = "Admin")]
-        [HttpDelete("{categoryId}")] // endpoint: /categories/:category_id
-        public async Task<bool> DeleteCategoryByIdAsync([FromRoute] Guid categoryId)
+        [HttpDelete("{categoryId}")]
+        public async Task<ActionResult<bool>> DeleteCategoryByIdAsync([FromRoute] Guid categoryId)
         {
-            return await _categoryService.DeleteCategoryByIdAsync(categoryId);
+            var foundCategory = await GetCategoryByIdAsync(categoryId);
+            if (foundCategory is null)
+            {
+                return NotFound("Category not found");
+            }
+
+            var result = await _categoryService.DeleteCategoryByIdAsync(categoryId);
+            return Ok(result);
         }
+        #endregion
     }
 }
